+++ src/main/java/org/iac2/service/architecturereconstruction/common/model/StructuralState.java
@@ -1,0 +1,7 @@
+package org.iac2.service.architecturereconstruction.common.model;
+
+public enum StructuralState {
+    EXPECTED,
+    NOT_EXPECTED,
+    REMOVED
+}
